Can now run program from batch files
Created new class with a 'main' function as an entry point to program.

Created a number of GenerateTags batch files that run the program based
upon the state of the database:
noDb - no db has been created
dbEmpty - db already exists but is empty
dbFull - db exists and is already filled with data

Simply double click the appropriate file and it will run.