ItemBasedCF Reconfigured for Better Metrics
Changed how the data is normalised in the configuration of a recommender
engine, which improved precision/recall.

Re-built database to consider only users with >=20 tags + cats. Minimum
number of ratings for a user is still the same at 14. Included dump.

Extended ItemBasedCF main to run metrics over multiple engine sizes and
multiple numbers of recommendations (separate the neighbourhood size and
total recommendations input with spaces when run, eg. Neighbourhood
Size: 10 20 30 40)