Inserting Mallet results into database
Created some batch files to run Mallet. At this stage, have to hardcode
a number of topics but I'm sure that is easily fixed.

Added a class to Data Generation (and a method to CapstoneDBConnection)
to insert Mallet stuff to database.

Completely abandoned and deleted the LDA project, which became redundant
after Java implementation was ditched in favour of batch files running
Mallet through cmd.

Batch files require the MALLET_HOME environment variable to be set
(looked into setting this programmatically but didn't have any luck).